requirements.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
elevenlabs
tavily-python
crewai
openai
numpy
pandas
sounddevice
soundfile
scipy
openai-whisper
